v1.4.2.2 minmax node (almost done) + response structures

+++ src/main/java/gameLogic/game/GameConstants.java
@@ -2,9 +2,9 @@ package gameLogic.game;
 
 public class GameConstants {
 
-    public static int FIGURE_NUMBER = 13;
-    public static int COLOR_NUMBER = 4;
-    public static int PLAYER_NUMBER = 4;
+    public static byte FIGURE_NUMBER = 13;
+    public static byte COLOR_NUMBER = 4;
+    public static byte PLAYER_NUMBER = 4;
     public static int DECK_SIZE = FIGURE_NUMBER * COLOR_NUMBER;
     public static int MAX_CARDS_PER_PLAYER = DECK_SIZE / PLAYER_NUMBER;
